refactor: 헤더 파일 관심사 분리
