+++ src/components/Button/CartButton.module.css
@@ -1,0 +1,18 @@
+.cartContainer {
+	position: relative;
+}
+
+.cartCount {
+	position: absolute;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	right: 0;
+	bottom: 0;
+	color: #000;
+	background: #fff;
+	width: 20px;
+	height: 20px;
+	border-radius: 50%;
+	font-weight: bold;
+}
