+++ src/components/Button/CartButton.tsx
@@ -1,0 +1,21 @@
+import S from "./CartButton.module.css";
+
+const CartButton = ({ count }: { count: number }) => {
+	return (
+		<div className={S.cartContainer}>
+			<svg xmlns="http://www.w3.org/2000/svg" width="33" height="32" viewBox="0 0 33 32" fill="none">
+				<path
+					d="M16.292 3.99998C15.4822 3.99998 14.7289 4.21152 14.0321 4.6346C13.3353 5.03844 12.7798 5.59613 12.3654 6.30767C11.97 7.01921 11.7722 7.78844 11.7722 8.61537V9.53844H7.30894L7.25244 10.4038L6.29199 28H26.292L25.3315 10.4038L25.275 9.53844H20.8118V8.61537C20.8118 7.78844 20.6046 7.01921 20.1903 6.30767C19.7948 5.59613 19.2487 5.03844 18.5519 4.6346C17.8551 4.21152 17.1018 3.99998 16.292 3.99998ZM16.292 5.84613C17.0453 5.84613 17.6856 6.11537 18.2129 6.65383C18.7402 7.19229 19.0039 7.84613 19.0039 8.61537V9.53844H13.5801V8.61537C13.5801 7.84613 13.8438 7.19229 14.3711 6.65383C14.8984 6.11537 15.5387 5.84613 16.292 5.84613ZM9.00386 11.3846H11.7722V14.1538H13.5801V11.3846H19.0039V14.1538H20.8118V11.3846H23.5801L24.3711 26.1538H8.2129L9.00386 11.3846Z"
+					fill="white"
+				/>
+			</svg>
+			{count !== 0 && (
+				<div className={S.cartCount}>
+					<p>{count}</p>
+				</div>
+			)}
+		</div>
+	);
+};
+
+export default CartButton;
