+++ src/components/Header/Header.module.css
@@ -10,22 +10,3 @@
 	background: #000;
 	color: white;
 }
-
-.cartContainer {
-	position: relative;
-}
-
-.cartCount {
-	position: absolute;
-	display: flex;
-	justify-content: center;
-	align-items: center;
-	right: 0;
-	bottom: 0;
-	color: #000;
-	background: #fff;
-	width: 20px;
-	height: 20px;
-	border-radius: 50%;
-	font-weight: bold;
-}
