+++ src/components/Toast/ErrorToast.module.css
@@ -1,8 +1,7 @@
 .toastContainer {
-	position: absolute;
-	bottom: 0;
-	left: 0;
-	width: 100%;
+	position: fixed;
+	top: 0;
+	width: 430px;
 	padding: 12px 0px;
 	font-size: 12px;
 	font-weight: 500;
@@ -19,12 +18,12 @@
 	}
 
 	10% {
-		transform: translateY(100%);
+		transform: translateY(150%);
 		opacity: 1;
 	}
 
 	90% {
-		transform: translateY(100%);
+		transform: translateY(150%);
 		opacity: 1;
 	}
