+++ src/components/pages/ProductListPage.module.css
@@ -24,19 +24,6 @@
 	justify-content: space-between;
 }
 
-.dropdownContainer {
-	display: flex;
-	align-items: center;
-	justify-content: space-between;
-}
-
-.dropdownContainer > * {
-	padding: 8px;
-	border-radius: 8px;
-	border: 1px solid rgba(0, 0, 0, 0.1);
-	background: #fff;
-}
-
 .itemContainer {
 	display: grid;
 	grid-template-columns: repeat(2, 1fr);
